Refactored WarehouseEditor toolbar components into separate component
